// apps/bridge/codegen/model/deploymentFamily.ts
/**
 * Generated by orval v6.25.0 🍺
 * Do not edit manually.
 * API
 * API docs
 * OpenAPI spec version: 1.0
 */

export type DeploymentFamily = typeof DeploymentFamily[keyof typeof DeploymentFamily];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DeploymentFamily = {
  optimism: 'optimism',
  arbitrum: 'arbitrum',
} as const;