// apps/bridge/codegen/model/deploymentType.ts
/**
 * Generated by orval v6.22.1 🍺
 * Do not edit manually.
 * API
 * API docs
 * OpenAPI spec version: 1.0
 */

export type DeploymentType = typeof DeploymentType[keyof typeof DeploymentType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DeploymentType = {
  mainnet: 'mainnet',
  testnet: 'testnet',
  devnet: 'devnet',
} as const;